Small updates to WI scrapers
