+++ can_tools/scrapers/official/WI/__init__.py
@@ -1,2 +1,2 @@
-from can_tools.scrapers.official.WI.wi_state import WisconsinCounties
+from can_tools.scrapers.official.WI.wi_county import WisconsinCounties
 from can_tools.scrapers.official.WI.wi_state import WisconsinState
